large refactor - ES6 js modularisation, indents
