+++ app/static/css/awesomesauce.css
@@ -1,49 +1,48 @@
 body {
-  background-color: #dead65;
-  background-image:
-    linear-gradient(0deg, rgba(0, 0, 0, 0.1) 1px, transparent 1px),
+    background-color: #dead65;
+    background-image: linear-gradient(0deg, rgba(0, 0, 0, 0.1) 1px, transparent 1px),
     linear-gradient(90deg, rgba(0, 0, 0, 0.1) 1px, transparent 1px);
-  background-size: 30px 30px;
+    background-size: 30px 30px;
 }
 
 .card {
-  background-color: #eac784;
+    background-color: #eac784;
 }
 
 .btn-primary {
-  background-color: #00704A;
-  border-color: #00704A;
+    background-color: #00704A;
+    border-color: #00704A;
 }
 
 .btn-primary:hover {
-  background-color: #005a3a;
-  border-color: #005a3a;
+    background-color: #005a3a;
+    border-color: #005a3a;
 }
 
 .navbar {
-  background: linear-gradient(90deg, #eac784 0%, #362415 100%);
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    background: linear-gradient(90deg, #eac784 0%, #362415 100%);
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
 }
 
 .navbar-light .navbar-nav .nav-link {
-  color: #fff;
-  transition: color 0.3s;
+    color: #fff;
+    transition: color 0.3s;
 }
 
 .navbar-light .navbar-nav .nav-link:hover {
-  color: #c79d69;
+    color: #c79d69;
 }
 
 
 .navbar-wrapper {
-  background-color: #f8f9fa; /* The same background color as the navbar */
+    background-color: #f8f9fa; /* The same background color as the navbar */
 }
 
 .navbar-container {
-  max-width: 400px; /* Adjust this value to change the width of the navbar */
-  margin: 0 auto;
+    max-width: 400px; /* Adjust this value to change the width of the navbar */
+    margin: 0 auto;
 }
 
 .container {
-  max-width: 800px; /* Adjust this value to change the width of the other container elements */
+    max-width: 800px; /* Adjust this value to change the width of the other container elements */
 }
