+++ app/static/js/modules/checkout.js
@@ -1,0 +1,5 @@
+export {handleCheckout};
+
+function handleCheckout() {
+    // stuff for checkout
+}
